Nest cordillera's Http2ServerConnection
`Http2ServerConnection` becomes `Http2.ServerConnection`, completing the `Http2.*`
family the second pass began. It was left out then because its file also held
`FlowWindow`; that is hoisted in the preceding commit.

`cordillera.Http2.scala` already opened with `import Http2.*` and `import
Http2.Error.Reason` — self-imports over the very object being extended, which the
donor file relied on too. They still resolve, so the new member needs no import
changes.
